Refactor package
* separate risk metrics and portfolio calculations
